* Stop warnings (for gcc 4.3.2 support)
* Stop using BOOST_FOREACH as for_each (conflicts with std library... just use BOOST_FOREACH)

All necessary for the upcoming v3 release.

 













git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@53017 16731396-06f5-0310-8873-f7f720988828

+++ public/dataclasses/I3TreeUtils.h
@@ -90,7 +90,7 @@ namespace I3TreeUtils
 	}
       }
       log_fatal("There is no object with ID = (%lli,%i) ",
-		major_id,
+		(long long)major_id,
 		minor_id);
       return *i;
     }
